requirements.txt
torchvision==0.3.0
opencv-python==4.2.0.34
Pillow
tensorboardX
matplotlib
tensorboard